bpf: tests: move IPv4 / IPv6 LB-no-backend tests into single file
We've started with proto-specific files, but are moving towards having
both proto families in a single test file. Makes it easier to see
commonalities.
